Fix printing of size of matrix and vectors in the compiler.
We would print non-casted unsigned char output, giving an an ascii
code on output, instead of printing the integer value of the size
field.

BUG=angle:539

Reviewed-on: https://chromium-review.googlesource.com/180650
Commit-Queue: Nicolas Capens <nicolascapens@chromium.org>
Tested-by: Nicolas Capens <nicolascapens@chromium.org>

+++ src/compiler/intermOut.cpp
@@ -44,9 +44,9 @@ TString TType::getCompleteString() const
     if (array)
         stream << "array[" << getArraySize() << "] of ";
     if (matrix)
-        stream << size << "X" << size << " matrix of ";
+        stream << static_cast<int>(size) << "X" << static_cast<int>(size) << " matrix of ";
     else if (size > 1)
-        stream << size << "-component vector of ";
+        stream << static_cast<int>(size) << "-component vector of ";
 
     stream << getBasicString();
     return stream.str();
